+++ name_map/variables.tf
@@ -37,7 +37,7 @@ variable "name_prefix_default" {
 variable "name_prepend_default" {
   type        = string
   default     = ""
-  description = "Prepended after key"
+  description = "Prepended before key"
 }
 
 variable "name_regex_allow_list" {